+++ Runtime/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("AchEngine.IAP.Tests")]
+[assembly: InternalsVisibleTo("AchEngine.Ads.Tests")]